refactor: rename 'migrations' module to 'authz_migration'
