+++ openedx_authz/handlers.py
@@ -15,8 +15,8 @@ from django.dispatch import receiver
 
 from openedx_authz.api.users import unassign_all_roles_from_user
 from openedx_authz.engine.utils import run_course_authoring_migration
+from openedx_authz.models.authz_migration import MigrationType, ScopeType
 from openedx_authz.models.core import ExtendedCasbinRule
-from openedx_authz.models.migrations import MigrationType, ScopeType
 from openedx_authz.models.subjects import UserSubject
 
 try:
